+++ frontend/src/components/calendar/TimeIndicator.tsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { CALENDAR_INDICATOR_COLOR, CELL_HEIGHT } from '../../helpers/styles'
 
+
 interface TimeIndicatorContainerProps {
     topOffset: number
 }
